backend/requirements.txt
fastapi~=0.135.3
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
asyncpg==0.31.0
pydantic-settings==2.13.1
alembic==1.18.4
httpx>=0.27,<1.0
pgvector>=0.4,<1.0